# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Salon SaaS API",
        "description": "API for salon booking, appointments, staff, and billing management",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/service-categories": {
            "get": {
                "tags": [
                    "Service Categories"
                ],
                "description": "Get list of service categories",
                "operationId": "getServiceCategories",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ServiceCategory"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Service Categories"
                ],
                "description": "Create a new service category",
                "operationId": "storeServiceCategory",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Hair Styling"
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "fa-cut",
                                        "nullable": true
                                    },
                                    "orderby": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": true,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServiceCategory"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/api/v1/service-categories/{service_category}": {
            "get": {
                "tags": [
                    "Service Categories"
                ],
                "description": "Get a single service category with its services",
                "operationId": "getServiceCategory",
                "parameters": [
                    {
                        "name": "service_category",
                        "in": "path",
                        "description": "Service Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServiceCategory"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Service Categories"
                ],
                "description": "Update a service category",
                "operationId": "updateServiceCategory",
                "parameters": [
                    {
                        "name": "service_category",
                        "in": "path",
                        "description": "Service Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Hair Care & Styling"
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "fa-scissors",
                                        "nullable": true
                                    },
                                    "orderby": {
                                        "type": "integer",
                                        "example": 2,
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": true,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServiceCategory"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Category not found"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Service Categories"
                ],
                "description": "Delete a service category (hard delete)",
                "operationId": "deleteServiceCategory",
                "parameters": [
                    {
                        "name": "service_category",
                        "in": "path",
                        "description": "Service Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Service category deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ServiceCategory": {
                "title": "ServiceCategory",
                "description": "ServiceCategory model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Hair Care"
                    },
                    "slug": {
                        "type": "string",
                        "example": "hair-care"
                    },
                    "icon": {
                        "type": "string",
                        "example": "fa-cut",
                        "nullable": true
                    },
                    "orderby": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Sanctum token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "cookieAuth": {
                "type": "apiKey",
                "description": "Laravel Sanctum session cookie authentication.",
                "name": "laravel_session",
                "in": "cookie"
            }
        }
    },
...
}